GOT 82% WOW

+++ database/create_insert_tables.php
@@ -1,5 +1,4 @@
 <?php
-
 //Creates a table called products
 $create_table_products =
 				"CREATE TABLE IF NOT EXISTS products (
@@ -8,16 +7,14 @@ $create_table_products =
 				product_quantity MEDIUMINT NOT NULL,
 				product_description VARCHAR(500) NOT NULL,
 				product_category VARCHAR(15),
-				product_price MEDIUMINT NOT NULL,
+				product_price DECIMAL NOT NULL,
 
 				PRIMARY KEY(ID)
 													  )";
-
 if(!$result = $database->query($create_table_products)){
     die('Could not create product table [' . $database->error . ']');
 	$result->close();
 }
-
 $create_table_address =
 				"CREATE TABLE IF NOT EXISTS address (
 				id MEDIUMINT NOT NULL AUTO_INCREMENT,
@@ -31,11 +28,8 @@ $create_table_address =
 
 				PRIMARY KEY(ID)
 														)";
-
 if(!$resultt = $database->query($create_table_address)){
 		die('Could not create product table [' . $database->error . ']');
 	$result->close();
 }
-
-
 ?>
